--- Application/Sources/Router/Route/Application/ApplicationRoute.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-enum ApplicationRoute: ScreenProtocol {
-    case application
-
-    var embedInNavigationView: Bool {
-        return true
-    }
-}
