--- Application/Sources/Router/Route/Entire/EntireRoute.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-enum EntireRoute: ScreenProtocol {
-    case entire
-    case myPage
-    case pointHistory
-    case bugReport
-
-    var embedInNavigationView: Bool {
-        return true
-    }
-}
