--- Application/Sources/Router/Route/Feed/FeedRoute.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-enum FeedRoute: ScreenProtocol {
-    case feed
-
-    var embedInNavigationView: Bool {
-        return true
-    }
-}
