--- Application/Sources/Router/Route/Home/HomeRoute.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-enum HomeRoute: ScreenProtocol {
-    case home
-    case mealDetail
-    case notification
-    case outingPass
-
-    var embedInNavigationView: Bool {
-        return true
-    }
-}
