+++ Application/Sources/Router/Route/Init/Router.swift
@@ -32,6 +32,8 @@ struct RouterContext<ScreenType: ScreenProtocol> {
 
 class Router<ScreenType, Factory: RouterFactory>: ObservableObject, RouterObject where Factory.Screen == ScreenType {
     @Published private var stack: [RouterContext<ScreenType>] = []
+    @Published var tabBarSelection: Int = 0
+
     var factory: Factory
 
     init(rootScreen: ScreenType, presentationType: PresentationType = .push, factory: Factory) {
