--- Application/Sources/Router/Route/Onboarding/OnboardingRoute.swift
@@ -1,13 +1,0 @@
-import Foundation
-
-enum OnboardingRoute: ScreenProtocol {
-    case launchScreen
-    case login
-    case signup
-    case onboarding
-    case main
-
-    var embedInNavigationView: Bool {
-        return false
-    }
-}
