--- Application/Sources/Router/Route/Router/XquareRouter.swift
@@ -1,8 +1,0 @@
-import Foundation
-
-typealias OnboardingRouter = Router<OnboardingRoute, OnboardingRouterFactory>
-typealias HomeRouter = Router<HomeRoute, HomeRouterFactory>
-typealias ScheduleRouter = Router<ScheduleRoute, ScheduleRouterFactory>
-typealias FeedRouter = Router<FeedRoute, FeedRouterFactory>
-typealias ApplicationRouter = Router<ApplicationRoute, ApplicationRouterFactory>
-typealias EntireRouter = Router<EntireRoute, EntireRouterFactory>
