--- Application/Sources/Router/Route/Schedule/ScheduleRoute.swift
@@ -1,10 +1,0 @@
-import Foundation
-
-enum ScheduleRoute: ScreenProtocol {
-    case schedule
-    case writeSchedule
-
-    var embedInNavigationView: Bool {
-        return true
-    }
-}
