+++ Application/Sources/Router/Route/XQuare/XquareRouter.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+typealias XquareRouter = Router<XquareRoute, XquareRouterFactory>
